requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
cryptography==42.0.8
ctransformers==0.2.27
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==4.4.2
deepdiff==7.0.1
deepspeech==0.9.3
defusedxml==0.7.1
Deprecated==1.2.14
diagrams==0.23.4
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docx==0.2.4
docx2txt==0.8
easyocr==1.7.1
email_validator==2.1.1
emoji==2.12.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.14.0
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
google-auth==2.30.0
googleapis-common-protos==1.63.1
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.7
ImageHash==4.3.1
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
InstructorEmbedding==1.0.1
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kubernetes==30.1.0
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-huggingface==0.0.3
langchain-mongodb==0.1.6
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.76
lazy_loader==0.4
llama-index==0.10.44
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.44
llama-index-embeddings-huggingface==0.2.1
llama-index-embeddings-instructor==0.1.3
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-chroma==0.1.9
llama-index-vector-stores-mongodb==0.1.5
llama-parse==0.4.4
llamaindex-py-client==0.1.19
looseversion==1.3.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.0.1
mistune==3.0.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.33.0
opencv-python-headless==4.10.0.82
openpyxl==3.1.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pipreqs==0.5.0
platformdirs==4.2.2
posthog==3.5.0
proglog==0.1.10
prompt_toolkit==3.0.46
protobuf==4.25.3
prov==2.0.1
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydot==2.0.0
pydub==0.25.1
Pygments==2.18.0
pymongo==4.7.3
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.3
rdflib==6.3.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.21.0
sentence-transformers==2.7.0
setuptools==69.5.1
shapely==2.0.4
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.30
srt==3.5.3
stack-data==0.6.2
starlette==0.37.2
striprtf==0.0.26
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
typed-ast==1.5.5
typer==0.12.3
typing_extensions==4.12.1
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.5
unstructured-client==0.23.3
urllib3==2.2.1
utils==1.0.2
uvicorn==0.30.1
vosk==0.3.45
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
llama-cpp-python